more dylib stuff
